add file for page html suppliers, categories, products, homepage and added some modif

+++ e-commerce/src/main/java/fr/utbm/ecommerce/service/ProductService.java
@@ -1,9 +1,9 @@
 package fr.utbm.ecommerce.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,22 +16,52 @@ public class ProductService {
 	@Autowired
 	private ProductDao productDao;
 
-	public Product addProduct(Product product) {
-		return productDao.save(product);
+	public boolean addProduct(Product product) {
+		boolean existed = productDao.existsById(product.getProductID());
+		if(existed == false){
+			productDao.save(product);
+			existed = true;
+		}else{
+			existed = false;
+		}
+		return existed;
 	}
 
-	public Product updateProduct(Product product) {
-		return productDao.save(product);
+	public boolean updateProduct(Product product) {
+		boolean existed = productDao.existsById(product.getProductID());
+		if(existed == true){
+			productDao.save(product);
+			existed = true;
+		}else{
+			existed = false;
+		}
+		return existed;
 	}
 
-	public void deleteProduct(Product product) {
-		productDao.delete(product);
+	public boolean deleteProduct(Product product) {
+		boolean existed = productDao.existsById(product.getProductID());
+		if(existed == true){
+			productDao.delete(product);
+		}
+        return existed;
 	}
 
 	public List<Product> getAllProduct() {
 		return (List<Product>) productDao.findAll();
 	}
+    
+    public Product getProductByName(String name){
+        Product product = productDao.getProductByName("%"+name+"%");
+        return product;
+    }
+    
+    public Product getProductById(int id){
+        return productDao.findById(id);
+    }
 
+    public boolean existed(int productid){
+    	return productDao.existsById(productid);
+    }
 //	public List<Product> getAllProductByName(@Param("name") String name) {
 //		return productDao.getAllProductByName("%" + name + "%");
 //	}
